Add initial configuration for DCLG planning portal
The TNA timestamp is the latest one at the moment; it may be updated to a
later crawl if one is done nearer their transition date.

Query string analysis hasn't been done yet - we need to get a list of URLs
before we can do that. Significant query params can be added later if we find
any.

The homepage will need to redirect to somewhere planning-specific on GOV.UK -
the content is still being worked on and a decision hasn't yet been reached
about where that should be, so this sets it to the department's homepage for
now.